[tests] Shift most integration tests into e2e tests
Trying to eventually empty it, so that we can remove the dependency on a WASM runtime just for tests
