+++ tests/BuiltinStringTest.sam
@@ -1,0 +1,10 @@
+class BuiltinStringTest {
+  function run(): unit = {
+    let value = Str.fromInt("42".toInt()) :: "!";
+    let _ = Process.println(value);
+    let a1 = "a";
+    let a2 = "a";
+    let _ = if a1 == a2 then {  } else { Process.panic("BAD") };
+    Process.println("Hello " :: "World!")
+  }
+}
