+++ tests/ConstantFoldingTest.sam
@@ -31,6 +31,9 @@ class ConstantFoldingTest {
       ConstantFoldingTest.test3(acc + 1 + 2 * 3 - 4 / 5 % 10000000 / 1234, i + 1)
     }
 
+  function test4(): int =
+    1 + 34 * 34 + 4 + 1 + 1231 / 28 + 100 / 100 - 2000 * 2000 / (10 * 10 * 10 * 4000)
+
   function run(): unit = {
     let _ = ForTests.assertIntEquals(
       /* expected */ 361200,
@@ -41,5 +44,6 @@ class ConstantFoldingTest {
     );
     let _ = ForTests.assertIntEquals(7000, ConstantFoldingTest.test2(0, 0));
     let _ = ForTests.assertIntEquals(1400, ConstantFoldingTest.test3(0, 0));
+    let _ = ForTests.assertIntEquals(1205, ConstantFoldingTest.test4());
   }
 }
