+++ tests/snapshot.txt
@@ -5,7 +5,13 @@ Test Name: Benchmark
 ========================================
 Test Name: BlockInIfElse
 ========================================
+Test Name: BuiltinStringTest
+42!
+Hello World!
+========================================
 Test Name: ConstantFoldingTest
+========================================
+Test Name: CorrectOp
 ========================================
 Test Name: CSETest
 ========================================
@@ -15,6 +21,8 @@ devmeggovibez
 Test Name: DifferentExpressionDemo
 ========================================
 Test Name: DifferentModulesDemo
+========================================
+Test Name: Empty
 ========================================
 Test Name: EvaluationOrder
 0
@@ -27,8 +35,26 @@ Test Name: EvaluationOrder
 7
 8
 9
+10
+11
+12
+13
+14
+15
+16
+17
+18
+19
+20
+21
+22
+23
+24
+25
 ========================================
 Test Name: GC
+========================================
+Test Name: IfElseTests
 ========================================
 Test Name: IntToString
 0
@@ -55,9 +81,45 @@ Test Name: MathFunctions
 Test Name: PrintHelloWorld
 Hello World!
 ========================================
+Test Name: ShortCircuitAndOr
+0
+1
+false
+0
+1
+true
+0
+false
+0
+false
+0
+true
+0
+true
+0
+1
+false
+0
+1
+true
+0
+1
+0
+1
+0
+0
+0
+0
+0
+1
+0
+1
+========================================
 Test Name: SortableList
 1
 2
 3
 4
+========================================
+Test Name: VariousSyntaxForms
 ========================================
